+++ engine/runtime/geometry/src/line_string/mod.rs
@@ -13,6 +13,8 @@ use crate::coordinate::CoordinateFrame;
 mod constructor;
 mod ops;
 #[cfg(feature = "new-geometry")]
+mod predicates;
+#[cfg(feature = "new-geometry")]
 mod validation;
 
 /// A polyline in 2D space, lying at a single optional elevation (2.5D).
