+++ engine/runtime/geometry/src/polygon_mesh/ops.rs
@@ -797,34 +797,6 @@ impl PolygonMesh3DData {
         Some(self.vertices()[i as usize][2])
     }
 }
-#[cfg(feature = "new-geometry")]
-impl crate::predicates::Equal for PolygonMesh2D {
-    fn equal(&self, rhs: &Self, tolerance: f64) -> crate::predicates::Result<bool> {
-        use crate::predicates::equal::surface_curves_2d;
-
-        crate::predicates::require_same_frame(self.frame(), rhs.frame())?;
-        Ok(surface_curves_2d(self)?.within(&surface_curves_2d(rhs)?, tolerance))
-    }
-}
-
-#[cfg(feature = "new-geometry")]
-impl crate::predicates::Equal for PolygonMesh3D {
-    fn equal(&self, _rhs: &Self, _tolerance: f64) -> crate::predicates::Result<bool> {
-        // TODO: replace with a topology-blind comparison — neither the triangles
-        // nor the faces are believed. For every vertex of either surface, find
-        // the closest point of the other (a point of the point set, so triangle
-        // interiors count), through an rstar prefilter over the triangle boxes
-        // and an exact point-to-triangle distance after it, returning as soon as
-        // one vertex is further than the tolerance; then match the boundary
-        // loops, if any, as 3D curves. Note vertex sampling alone is unsound on
-        // closed surfaces: a cube and the same cube with a tunnel drilled
-        // through it have every vertex of each lying on the other.
-        Err(crate::predicates::PredicateError::Unsupported {
-            geometry: "PolygonMesh3D",
-        })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
