+++ engine/runtime/geometry/src/solid/ops.rs
@@ -328,22 +328,6 @@ impl Elevation for Solid {
     }
 }
 
-#[cfg(feature = "new-geometry")]
-impl crate::predicates::Equal for Solid {
-    fn equal(&self, _rhs: &Self, _tolerance: f64) -> crate::predicates::Result<bool> {
-        // TODO: replace with a topology-blind comparison — neither the triangles
-        // nor the faces are believed. For every vertex of either surface, find
-        // the closest point of the other (a point of the point set, so triangle
-        // interiors count), through an rstar prefilter over the triangle boxes
-        // and an exact point-to-triangle distance after it, returning as soon as
-        // one vertex is further than the tolerance; then match the boundary
-        // loops, if any, as 3D curves. Note vertex sampling alone is unsound on
-        // closed surfaces: a cube and the same cube with a tunnel drilled
-        // through it have every vertex of each lying on the other.
-        Err(crate::predicates::PredicateError::Unsupported { geometry: "Solid" })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
