+++ engine/runtime/geometry/src/triangular_mesh/mod.rs
@@ -20,6 +20,8 @@ mod constructor;
 mod feature_write;
 mod ops;
 #[cfg(feature = "new-geometry")]
+mod predicates;
+#[cfg(feature = "new-geometry")]
 mod validation;
 
 /// A triangle mesh in 2D space, lying at a single optional elevation.
